requirements.txt
PyMuPDF
pytesseract
Pillow
pandas
openpyxl
langchain_community
langchain-gigachat
gigachat